requirements.txt
# requirements.txt

# Основная библиотека для создания Telegram-бота
python-telegram-bot==21.0.1

# Для загрузки переменных окружения из .env файла
python-dotenv

# ORM для асинхронной работы с PostgreSQL и драйвер
sqlalchemy[asyncpg]

# Для миграций схемы базы данных
alembic

# Официальный SDK для Google Gemini API
google-generativeai

# Для управления конфигурацией через .env
pydantic-settings